2D simulation of clock hands
Using the idea from Dynamics of Structures: Networks of Springs and Dashpots with Point Masses at Nodes, I built amodel of two clock hands spinning around the clock center. By implementing different values of mass, initial velocity, dashpot constant, and spring constant, as well as the magnitude of the time step, I aim to unravel the effects of the physical variables on the spinning motion.